+++ frontend/src/components/Calendar/Core/index.tsx
@@ -1,6 +1,5 @@
 import { Flex } from '@/components/Flex';
 
-import { CalendarHeader } from './CalendarHeader';
 import { coreStyle } from './index.css';
 import { SelectedWeek } from './SelectedWeek';
 import { TimeControl } from './TimeControl';
@@ -9,10 +8,10 @@ export const Core = () => (
   <Flex
     className={coreStyle}
     direction='column'
+    height='8.375rem'
     width='100%'
   >  
     <TimeControl />
     <SelectedWeek />
-    <CalendarHeader />
   </Flex>
 );
