+++ frontend/src/features/my-calendar/ui/MyCalendar/index.tsx
@@ -50,6 +50,7 @@ export const MyCalendar = () => {
   return (
     <Calendar {...calendar} className={calendarStyle}>
       <Calendar.Core />
+      <Calendar.Header />
       <CalendarTable />
     </Calendar>
   );
